+++ CHANGELOG.md
@@ -13,6 +13,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Adding the mermaid figure to the html export 
 - ODCS export: Export physical type if the physical type is configured in config object
 
+### Added
+
+- `datacontract import --format spark`: Added support for spark importer table level comments (#761)
+- `datacontract import` respects `--owner` and `--id` flags (#753)
+
+### Fixed
+
+- `datacontract export --format sodacl`: Fix resolving server when using `--server` flag (#768)
+
 ## [0.10.26] - 2025-05-16
 
 ### Changed