+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Changed
+=======
 ### Added
 
 - `datacontract import --format spark`: Added support for spark importer table level comments (#761)
@@ -15,6 +17,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - `datacontract export --format sodacl`: Fix resolving server when using `--server` flag (#768)
+- `datacontract export --format dbt`: Fixed DBT export behaviour of constraints to default to data tests when no model type is specified in the datacontract model
+
 
 ## [0.10.26] - 2025-05-16